feat: make AirPlay route metadata available
This commit slightly rewires the data stream from Swift to React Native to
not only return a boolean indicating whether AirPlay is active, but
rather relay the entire set of routes so that they can be used in
interfaces.

A new hook is introduced called `useAirPlayRoutes()` that makes this
information available. The existing routes are modified to fit the new
format. Functionality should remain unchanged.

Additionally, the example app and prettierrc are upgraded to the latest
versions.